Moved CompressionUtils to androidshared
