+++ collect_app/src/androidTest/java/org/odk/collect/android/support/StubBarcodeViewDecoder.kt
@@ -7,7 +7,7 @@ import com.google.zxing.Result
 import com.journeyapps.barcodescanner.BarcodeResult
 import com.journeyapps.barcodescanner.DecoratedBarcodeView
 import org.odk.collect.android.views.BarcodeViewDecoder
-import org.odk.collect.shared.CompressionUtils
+import org.odk.collect.androidshared.utils.CompressionUtils
 import java.io.IOException
 
 class StubBarcodeViewDecoder : BarcodeViewDecoder() {
