+++ collect_app/src/main/java/org/odk/collect/android/configure/qr/QRCodeScannerFragment.kt
@@ -13,8 +13,8 @@ import org.odk.collect.android.injection.DaggerUtils
 import org.odk.collect.android.projects.CurrentProjectProvider
 import org.odk.collect.android.storage.StoragePathProvider
 import org.odk.collect.androidshared.ui.ToastUtils.showLongToast
+import org.odk.collect.androidshared.utils.CompressionUtils
 import org.odk.collect.settings.ODKAppSettingsImporter
-import org.odk.collect.shared.CompressionUtils
 import java.io.File
 import java.io.IOException
 import java.util.zip.DataFormatException
